refactor(materializer): extract conversation-rebuild + reuse paths (file #3)
Lift the full-rebuild path (materializeConversation/WithOperationIndex) and the
cache-reuse predicates into conversation-rebuild.ts; the dispatcher imports them.
Materializer core now = dispatcher + scene assembly + public spine. Net 74/74, check 0.
